requirements.txt
datasets
matplotlib
numpy
torch
torchvision
torchtext
torchaudio
tqdm
json5
scikit-learn
opencv-python
efficientnet_pytorch
timm
pillow
zipfile36
grad-cam
transformers
streamlit
huggingface_hub
scikit-image
sentencepiece
torchmetrics
